Split the bundle file
